+++ src/cmd/fetch.rs
@@ -59,7 +59,7 @@ pub async fn fetch_urls(
     target_reader.read(&mut target_bookmarks)?;
 
     for url in urls {
-        let mut bookmark = TargetBookmark::new(url, now, None, HashSet::new());
+        let mut bookmark = TargetBookmark::new(url, now, None, HashSet::new(), HashSet::new());
         fetch_and_add(client, cache, &mut bookmark, true).await?;
         info!("Fetched website for {url}");
         target_bookmarks.insert(bookmark);
@@ -134,8 +134,6 @@ async fn fetch_and_add(
 
                 if let Err(err) = cache.replace(html, bookmark).await {
                     warn!("Can't replace website ({}) in cache: {}", bookmark.url, err);
-                } else {
-                    bookmark.last_cached = Some(Utc::now().timestamp_millis());
                 }
             }
             Ok(None) => (),
@@ -151,8 +149,6 @@ async fn fetch_and_add(
 
                 if let Err(err) = cache.add(html, bookmark).await {
                     warn!("Can't add website ({}) to cache: {}", bookmark.url, err);
-                } else {
-                    bookmark.last_cached = Some(Utc::now().timestamp_millis());
                 }
             }
             Ok(None) => (),
@@ -177,7 +173,7 @@ pub async fn fetch_diff(config: &Config, args: FetchArgs) -> Result<(), anyhow::
     target_reader.read(&mut target_bookmarks)?;
 
     for url in args.diff {
-        let bookmark = target_bookmarks.get(&url);
+        let bookmark = target_bookmarks.get_mut(&url);
 
         if let Some(bookmark) = bookmark {
             if let Some(cached_website_before) = cache.get(bookmark)? {
@@ -235,6 +231,7 @@ mod tests {
                     last_imported: now.timestamp_millis(),
                     last_cached: None,
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
             (
@@ -245,6 +242,7 @@ mod tests {
                     last_imported: now.timestamp_millis(),
                     last_cached: None,
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
         ]));
@@ -296,6 +294,7 @@ mod tests {
                     last_imported: now.timestamp_millis(),
                     last_cached: None,
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
             (
@@ -306,6 +305,7 @@ mod tests {
                     last_imported: now.timestamp_millis(),
                     last_cached: None,
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
         ]));
@@ -357,6 +357,7 @@ mod tests {
                     last_imported: now,
                     last_cached: Some(now),
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
             (
@@ -367,6 +368,7 @@ mod tests {
                     last_imported: now,
                     last_cached: None,
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
         ]));
@@ -383,7 +385,7 @@ mod tests {
             .add(
                 "<html><head></head><body><p>Test content (already cached)</p></body></html>"
                     .to_owned(),
-                target_bookmarks.get("https://test_url1.com").unwrap(),
+                target_bookmarks.get_mut("https://test_url1.com").unwrap(),
             )
             .await
             .unwrap();
@@ -428,6 +430,7 @@ mod tests {
                     last_imported: now,
                     last_cached: Some(now),
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
             (
@@ -438,6 +441,7 @@ mod tests {
                     last_imported: now,
                     last_cached: None,
                     sources: HashSet::new(),
+                    cache_modes: HashSet::new(),
                 },
             ),
         ]));
@@ -454,7 +458,7 @@ mod tests {
             .add(
                 "<html><head></head><body><p>Test content (already cached)</p></body></html>"
                     .to_owned(),
-                target_bookmarks.get("https://test_url1.com").unwrap(),
+                target_bookmarks.get_mut("https://test_url1.com").unwrap(),
             )
             .await
             .unwrap();